@@ -4,8 +4,10 @@
   "description": "Sauce Labs test runner helper libary.",
   "author": "devx <dev@saucelabs.com>",
   "license": "MIT",
-  "main": "./src/main.js",
+  "main": "./lib/index",
+  "typings": "./lib/index.d.ts",
   "scripts": {
+    "build": "tsc",
     "test": "run-s test:*",
     "test:lint": "eslint tests/ src/",
     "test:unit": "jest --env node",
@@ -18,30 +20,37 @@
   "keywords": [],
   "dependencies": {
     "lodash": "^4.17.20",
-    "npm": "^6.14.9"
+    "npm": "^6.14.15"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
     "@types/chai": "^4.2.12",
-    "@types/jest": "^26.0.19",
+    "@types/jest": "^27.0.3",
     "@types/jquery": "^3.5.5",
+    "@types/lodash": "^4.14.177",
     "@types/mocha": "^8.2.0",
+    "@types/mock-fs": "^4.13.1",
     "@types/node": "^14.14.13",
+    "@types/npm": "^2.0.32",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "axios": "^0.21.1",
     "bluebird": "^3.7.2",
     "eslint": "^7.15.0",
     "eslint-config-appium": "^4.5.0",
     "eslint-plugin-jest": "^24.1.3",
     "husky": "^4.3.6",
-    "jest": "^26.6.3",
+    "jest": "^27.3.1",
     "mocha": "^8.2.1",
     "mock-fs": "^4.13.0",
     "npm-run-all": "^4.1.5",
     "release-it": "^14.4.1",
-    "typescript": "^3.9.7"
+    "ts-jest": "^27.0.7",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.2"
   },
   "files": [
-    "src"
+    "lib",
+    "LICENSE"
   ],
   "husky": {
     "hooks": {
